refactor OAuth code and related tests
Decoupled OAuth protocol related code from UI aspects by introducing OAuthFlowUtil.java and OAuthSecretFlowUtil.java. Renamed related tests.
