+++ src/main/java/com/salesforce/dataloader/ui/OAuthTokenFlow.java
@@ -27,16 +27,15 @@
 package com.salesforce.dataloader.ui;
 
 import com.salesforce.dataloader.config.Config;
-import com.salesforce.dataloader.model.OAuthToken;
+import com.salesforce.dataloader.oauth.OAuthFlowUtil;
+
 import org.eclipse.swt.browser.Browser;
 import org.eclipse.swt.browser.ProgressEvent;
 import org.eclipse.swt.widgets.Shell;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URISyntaxException;
-import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
-import java.util.Map;
+
 
 /**
  * the oauth token flow. this is normally used for client to server where the client is not a secured environment
@@ -55,14 +54,7 @@ public class OAuthTokenFlow extends OAuthFlow {
 
     @Override
     public String getStartUrl(Config config) throws UnsupportedEncodingException {
-        return getStartUrlImpl(config);
-    }
-
-    public static String getStartUrlImpl(Config config) throws UnsupportedEncodingException {
-        return config.getString(Config.OAUTH_SERVER) +
-                "/services/oauth2/authorize?response_type=token&display=popup&client_id=" +
-                config.getString(Config.OAUTH_CLIENTID) + "&redirect_uri=" +
-                URLEncoder.encode(config.getString(Config.OAUTH_REDIRECTURI), StandardCharsets.UTF_8.name());
+        return OAuthFlowUtil.getStartUrlImpl(config);
     }
 
     public static class OAuthTokenBrowserLister extends OAuthBrowserListener {
@@ -79,7 +71,7 @@ public class OAuthTokenFlow extends OAuthFlow {
         public void completed(ProgressEvent progressEvent) {
             super.completed(progressEvent);
             try {
-                boolean handled = handleCompletedUrl(browser.getUrl(), config);
+                boolean handled = OAuthFlowUtil.handleCompletedUrl(browser.getUrl(), config);
                 if (handled) {
                     setResult(true);
                     shell.close();
@@ -89,50 +81,5 @@ public class OAuthTokenFlow extends OAuthFlow {
                 doSimpleErrorHandling(browser.getUrl(), e, logger);
             }
         }
-
-        public static boolean handleCompletedUrl(String url, Config config) throws URISyntaxException {
-            Map<String, String> params = getQueryParameters(url);
-
-            if (params.containsKey("access_token")){
-                //we don't use most of this but I still like to track what we should get
-                OAuthToken token = new OAuthToken();
-                token.setInstanceUrl(params.get("instance_url"));
-                token.setId(params.get("id"));
-                token.setAccessToken(params.get("access_token"));
-
-                //optional parameters
-                if (params.containsKey("refresh_token")) {
-                    token.setRefreshToken(params.get("refresh_token"));
-                }
-
-                //currently unused parameters
-                if (params.containsKey("scope")) {
-                    token.setScope(params.get("scope"));
-                }
-                if (params.containsKey("signature")) {
-                    token.setSignature(params.get("signature"));
-                }
-                if (params.containsKey("token_type")) {
-                    token.setTokenType(params.get("token_type"));
-                }
-                if (params.containsKey("issued_at")) {
-                    String issued_at = params.get("issued_at");
-                    if (issued_at != null && !issued_at.equals("")) {
-                        token.setIssuedAt(Long.valueOf(issued_at));
-                    }
-                }
-
-
-                config.setValue(Config.OAUTH_ACCESSTOKEN, token.getAccessToken());
-                config.setValue(Config.OAUTH_REFRESHTOKEN, token.getRefreshToken());
-                config.setValue(Config.ENDPOINT, token.getInstanceUrl());
-
-                return true;
-            }
-
-            return false;
-        }
     }
-
-
 }
